requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery==5.5.3
redis==6.1.1
python-decouple==3.8
django-filter==24.3
gunicorn==23.0.0
pytest==8.3.2 

psycopg2-binary==2.9.10
drf-spectacular==0.27.1 